refactor(maps): extract useTokenInteractions hook from ActiveMapStage
Move token selection, layer-move, label toggle, GM-gated removal + the
delete-confirm/Esc keyboard handling, the right-click context menu state, and
the visibleTokens/tokenCounts/canMoveToken derivations into a useTokenInteractions
hook. Token dragging stays in the stage (it shares the dragRef) and reads
selectedTokenIds from the hook for group moves. Destructured into the same local
names so the render/dialog/menu wiring is unchanged. Verified by the
monster-tokens + token group-move e2e specs (8 tests).
